Use policy to control access to User model records

+++ app/Http/Controllers/Cms/UsersController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\UserStoreRequest;
 use App\Http\Requests\UserUpdateRequest;
 use App\Models\User;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Validator;
@@ -13,11 +14,16 @@ use Inertia\Inertia;
 
 class UsersController extends Controller
 {
+    use AuthorizesRequests;
+
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
+
+        $this->authorize('viewAny', User::class);
+
         return Inertia::render('Resources/Index', [
             'title' => 'Users',
             'resourceName' => 'users',
@@ -35,6 +41,9 @@ class UsersController extends Controller
      */
     public function create()
     {
+
+        $this->authorize('create', User::class);
+
         return Inertia::render('Resources/Create', [
             'title' => 'Users > Add User',
             'schema' => json_decode(User::$createSchema),
@@ -49,6 +58,8 @@ class UsersController extends Controller
      */
     public function store(UserStoreRequest $request): JsonResponse
     {
+        $this->authorize('create', User::class);
+
         // extract and validate metadata from the json field
         $data = $this->_validatedJsonMetadata($request);
 
@@ -69,6 +80,8 @@ class UsersController extends Controller
      */
     public function edit(User $user)
     {
+        $this->authorize('update', $user);
+
         return Inertia::render('Resources/Edit', [
             'title' => 'Users > Edit User',
             'schema' => json_decode(User::$editSchema),
@@ -87,6 +100,9 @@ class UsersController extends Controller
      */
     public function update(UserUpdateRequest $request, User $user): JsonResponse
     {
+
+        $this->authorize('update', $user);
+
         // extract and validate metadata from the json field
         $data = $this->_validatedJsonMetadata($request, $user);
 
@@ -106,6 +122,9 @@ class UsersController extends Controller
      */
     public function destroy(User $user): JsonResponse
     {
+
+        $this->authorize('delete', User::class);
+
         $response = $user->delete();
 
         return $response
@@ -118,7 +137,7 @@ class UsersController extends Controller
      *
      * @return array
      */
-    private function _validatedJsonMetadata($request, $user)
+    private function _validatedJsonMetadata($request, $user = null)
     {
         // extract metadata from json field
         $metadata = $this->_extractMetadataFromJsonData($request->json);
